Scorekeeping working
The scorekeeping works when the monitor is moved inside the onClick
function. The round tracker going up twice seems to be a result
of the function actually running twice.

Oh! It must be re-applying the .one click limit thingy twice.
Maybe that's what's doing it?

Anyway, time to take on all the API stuff.