Fixed draw error
The draw conditions were not properly reseting after a
player victory. As such, after someone won, the game could no longer
detect a draw. Removed the extraneous winner detection conditions
and set it to simply declare a draw if the round > 8. This should
be fine since it is the last item on the if else chain and it appears
to be working correctly now.